Update README.md
This is my first repository and i have share a mini project based on client server architecture. The vehicle Service Center Management System (VSCM) is a progressive step in the field of service centers. Any car user can make use of such site to locate and communicate with the service centers in the vicinity. The proposed system can be used by any automobile user.